Apply estimator to collected matlab data
This is a first draft/attempt at fitting the data with the model. It is
likely that there is a mistake either in the model or the current
assumptions are inadequate in capturing the dynamics of the car. A LLS
estimator has been written as a comparison to PBGF and it seems to be
doing just as poorly.