Fixed-lag smoother class
This is the first-pass of writing the algorithm. It has been proven to
be functional based on a linear example which will be pushed in the
subsequent commit.

There are several improvements that can be made to the class:
- removal of redundant variables
- logic to enable asynchronous prediction & update. The current version
works in most cases except when multiple update step is invoked in a row.
- remove the use of stochastic clone as statistical dependency between
current and predicted state can be found by using a copy of the sampled
points. The stochastic clone is useful in verifying that the first
iteration of the backward pass is correct.

An implementation of a fully augmented fixed-lag smoother is encouraged
for its use in comparing against and debugging the current and future
implementations.